Added dry run capabilities

+++ rudaux/rudaux/course.py
@@ -34,7 +34,7 @@ class Course(object):
     Course object for managing a Canvas/JupyterHub/nbgrader course.
     """
 
-    def __init__(self, course_dir, allow_canvas_cache = False):
+    def __init__(self, course_dir, dry_run = False, allow_canvas_cache = False):
         """
         Initialize a course from a config file. 
         :param course_dir: The directory your course. If none, defaults to current working directory. 
@@ -45,6 +45,7 @@ class Course(object):
         """
 
         self.course_dir = course_dir
+        self.dry_run = dry_run
 
         #=======================================#
         #              Load Config              #
@@ -202,24 +203,33 @@ class Course(object):
     
     def save_jupyterhub_state(self):
         print('Saving the JupyterHub state...')
-        with open(self.jupyterhub_cache_filename, 'wb') as f:
-            pk.dump((self.submissions, self.snapshots), f)
-        print('Done.')
+        if not self.dry_run:
+            with open(self.jupyterhub_cache_filename, 'wb') as f:
+                pk.dump((self.submissions, self.snapshots), f)
+            print('Done.')
+        else:
+            print('[Dry Run: state not saved]')
         return
 
-    def jupyterhub_snapshot(self):
+    def jupyterhub_snapshot(self, dry_run = False):
         print('Taking snapshots')
         for a in self.assignments:
             if a.due_at and a.due_at < plm.now() and a.name not in self.snapshots:
                 print('Assignment ' + a.name + ' is past due and no snapshot exists yet. Taking a snapshot [' + a.name + ']')
-                self.jupyterhub.snapshot_all(a.name)
-                self.snapshots.append(a.name)
+                self.jupyterhub.snapshot_all(a.name, dry_run)
+                if not dry_run:
+                    self.snapshots.append(a.name)
+                else:
+                    print('[Dry Run: snapshot name not added to taken list; would have added ' + a.name + ']')
             for over in a.overrides:
                 snapname = a.name + '-override-' + over['id']
                 if over['due_at'] and over['due_at'] < plm.now() and not (snapname in self.snapshots):
                     print('Assignment ' + a.name + ' has override ' + over['id'] + ' for student ' + over['student_ids'][0] + ' and no snapshot exists yet. Taking a snapshot [' + snapname + ']')
-                    self.jupyterhub.snapshot_user(over['student_ids'][0], snapname)
-                    self.snapshots.append(snapname)
+                    self.jupyterhub.snapshot_user(over['student_ids'][0], snapname, dry_run)
+                    if not dry_run:
+                        self.snapshots.append(snapname)
+                    else:
+                        print('[Dry Run: snapshot name not added to taken list; would have added ' + snapname + ']')
         print('Done.')
         self.save_jupyterhub_state()
